[Fix] 카드뷰, 카드Board 뷰 모두 동일한 custom view 상속받고있었는데, 특정 부분에서 공통기능이 달라져서 객체를 나누고 상속형태로 변경
